requirements.txt
google-api-python-client
google-generativeai
google-ai-generativelanguage
flask
Gunicorn
